fix minor number cast mistake

+++ grails-app/services/com/unifina/service/SerializationService.groovy
@@ -44,7 +44,7 @@ class SerializationService {
 	}
 
 	int serializationMaxBytes() {
-		Integer v = MapTraversal.getLong(grailsApplication.getConfig(), MAX_BYTES_CONFIG_KEY)
+		Long v = MapTraversal.getLong(grailsApplication.getConfig(), MAX_BYTES_CONFIG_KEY)
 		Assert.notNull(v, "Missing key \"" + MAX_BYTES_CONFIG_KEY + "\" from grailsApplication configuration")
 		return v
 	}
